Fix build errors: Guid-to-string conversion and MagickGeometry uint cast
- Use photo.Id.ToString() for ThumbnailCache string keys
- Cast maxDimension to uint for MagickGeometry constructor

+++ src/PhotoCull/Services/LibRaw/RawPreviewExtractor.cs
@@ -129,7 +129,7 @@ public static class RawPreviewExtractor
         var maxSide = Math.Max((int)image.Width, (int)image.Height);
         if (maxSide > maxDimension)
         {
-            var geometry = new MagickGeometry(maxDimension, maxDimension)
+            var geometry = new MagickGeometry((uint)maxDimension, (uint)maxDimension)
             {
                 IgnoreAspectRatio = false
             };
